requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
selenium
webdriver-manager
beautifulsoup4
ics
arrow
Flask-Cors
